Added Impl of KeyFactorsData and KeyActivities

+++ src/main/java/com/acko/insuredassetcredibility/dao/ScoreDao.java
@@ -21,10 +21,10 @@ import java.util.Map;
 public class ScoreDao {
     @Id
     String assetId;
-    Double score;
-    Map<KeyFactors, Double> keyFactorScores;
+    Integer score;
+    Map<KeyFactors, Integer> keyFactorScores;
     Date refreshDate;
     List<KeyActivities> activitiesList;
-    List<KeyFactorsData>keyFactorsData;
+    List<KeyFactorsData> keyFactorsData;
 
 }
